Refactor: github, algorithm, grade 모듈을 stat 모듈로 합침
