--- src/algorithm/algorithm.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AlgorithmController } from './algorithm.controller';
-import { AlgorithmService } from './algorithm.service';
-import { AlgorithmRepository } from './algorithm.repository';
-
-@Module({
-    controllers: [AlgorithmController],
-    providers: [AlgorithmService, AlgorithmRepository],
-    exports: [AlgorithmService],
-})
-export class AlgorithmModule {}
