+++ src/total/total.service.spec.ts
@@ -2,9 +2,9 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { TotalService } from './total.service';
 import { TotalRepository } from './total.repository';
 import { User } from '../Entity/user';
-import { GithubService } from '../github/github.service';
-import { GradeService } from '../grade/grade.service';
-import { AlgorithmService } from '../algorithm/algorithm.service';
+import { GithubService } from '../stat/service/github.service';
+import { GradeService } from '../stat/service/grade.service';
+import { AlgorithmService } from '../stat/service/algorithm.service';
 
 const mockTotalRepository = {
     save: jest.fn(),
