+++ src/total/total.service.ts
@@ -1,9 +1,9 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { TotalRepository } from './total.repository';
 import { TotalPoint } from '../Entity/totalPoint';
-import { GithubService } from '../github/github.service';
-import { AlgorithmService } from '../algorithm/algorithm.service';
-import { GradeService } from '../grade/grade.service';
+import { GithubService } from '../stat/service/github.service';
+import { AlgorithmService } from '../stat/service/algorithm.service';
+import { GradeService } from '../stat/service/grade.service';
 import { Github } from '../Entity/github';
 import { Grade } from '../Entity/grade';
 import { Algorithm } from '../Entity/algorithm';
